Add transcoding for beancount
The transcode subcommand will transcode an entire knut ledger into a
format compatible with the beancount tool. This allows using
beancount's amazing tooling ecosystem for analysis, while managing the
journal in knut.

To work around the differences in implementation for cross-commodity
support between knut and beancount, transcode will convert all amounts
into a single commodity. Therefore, the --commodity / -c argument is
required, and prices for all commodities in the journal are
needed. A benefit of this conversion is that fava will correctly
account for all market value changes, which is difficult to do with a
pure beancount journal.